[refactor] 감상 기록 / 패키지 변경 (#71)
