+++ android/app/src/main/java/com/teamtripdraw/android/ui/post/writing/PostWritingViewModel.kt
@@ -1,4 +1,4 @@
-package com.teamtripdraw.android.ui.post
+package com.teamtripdraw.android.ui.post.writing
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
