+++ android/app/src/main/res/layout/activity_post_writing.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="postWritingViewModel"
-            type="com.teamtripdraw.android.ui.post.PostWritingViewModel" />
+            type="com.teamtripdraw.android.ui.post.writing.PostWritingViewModel" />
 
     </data>
